+++ pylib/anki/BUILD.bazel
@@ -4,6 +4,7 @@ load("@py_deps//:requirements.bzl", "requirement")
 load("@rules_python//experimental/python:wheel.bzl", "py_package", "py_wheel")
 load("//:defs.bzl", "anki_version")
 load("//pylib:orjson.bzl", "orjson_if_available")
+load("//pylib:protobuf.bzl", "py_proto")
 
 copy_file(
     name = "buildinfo",
@@ -32,7 +33,6 @@ py_library(
         "py.typed",
         ":buildinfo",
         ":hooks_gen",
-        ":proto_py",
         "//pylib/anki/_backend",
     ],
     imports = [
@@ -40,6 +40,7 @@ py_library(
     ],
     visibility = ["//visibility:public"],
     deps = [
+        ":proto",
         requirement("beautifulsoup4"),
         requirement("decorator"),
         requirement("distro"),
@@ -107,8 +108,6 @@ filegroup(
     ],
 )
 
-load("//pylib:protobuf.bzl", "py_proto")
-
 py_proto(
     name = "proto_py",
     srcs = ["//proto"],
@@ -117,10 +116,15 @@ py_proto(
     ],
 )
 
-# only used for genbackend.py
 py_library(
-    name = "proto_lib",
-    srcs = [":proto_py", "__init__.py"],
+    name = "proto",
+    srcs = [
+        "__init__.py",
+        # includes the .py files
+        ":proto_py",
+    ],
+    # includes the .pyi files
+    data = [":proto_py", "py.typed"],
     imports = [".."],
     visibility = ["//visibility:public"],
 )
