+++ pylib/tests/shared.py
@@ -6,7 +6,7 @@ import shutil
 import tempfile
 import time
 
-from anki import Collection as aopen
+from anki.collection import Collection as aopen
 
 # Between 2-4AM, shift the time back so test assumptions hold.
 lt = time.localtime()
