+++ qt/dmypy.py
@@ -30,8 +30,8 @@ if subprocess.run(
         "--",
         "--config-file",
         "qt/mypy.ini",
-        "bazel-bin/qt/dmypy.runfiles/anki/pylib/anki",
-        "bazel-bin/qt/dmypy.runfiles/anki/qt/aqt",
+        "bazel-bin/qt/dmypy.runfiles/ankidesktop/pylib/anki",
+        "bazel-bin/qt/dmypy.runfiles/ankidesktop/qt/aqt",
         "--python-executable",
         os.path.abspath("pip/stubs/extendsitepkgs"),
     ],
